refactor: extract the project info getting logic to a hook for reusability
